+++ packages/api/src/modules/user/user.service.test.ts
@@ -51,6 +51,7 @@ import { TokenValidationDtoPositiveResponse } from "dto";
 import { TokenValidationType } from "dto";
 import { AgentTypeEnum } from "dto";
 import { AgentJobTypeEnum, TerritorialScopeEnum } from "dto";
+import { removeSecrets } from "../../shared/helpers/RepositoryHelper";
 
 jest.useFakeTimers().setSystemTime(new Date("2022-01-01"));
 
@@ -213,6 +214,7 @@ describe("User Service", () => {
             // @ts-expect-error: unknown error
             mockedUserRepository.update.mockImplementation(() => ({
                 ...USER_WITHOUT_SECRET,
+                ...USER_ACTIVATION_INFO,
                 hashPassword: "qdqdqzdqzd&",
             }));
             mockGetUserById.mockImplementation(async id => UNACTIVATED_USER);
@@ -222,6 +224,14 @@ describe("User Service", () => {
         it("should call userRepository.update()", async () => {
             await userService.activate("token", USER_ACTIVATION_INFO);
             expect(userRepository.update).toHaveBeenCalledTimes(1);
+        });
+
+        it("should notify user updated", async () => {
+            await userService.activate("token", USER_ACTIVATION_INFO);
+            expect(notifyService.notify).toHaveBeenCalledWith(NotificationType.USER_UPDATED, {
+                ...USER_WITHOUT_SECRET,
+                ...USER_ACTIVATION_INFO,
+            });
         });
 
         it("should call validateAndSanitizeActivationUserInfo()", async () => {
@@ -1241,6 +1251,28 @@ describe("User Service", () => {
             )) as TokenValidationDtoPositiveResponse;
 
             expect(actual.type).toBe(TokenValidationType.FORGET_PASSWORD);
+        });
+    });
+
+    describe("getUserWithoutSecret", () => {
+        const EMAIL = "user@mail.fr";
+
+        it("gets user from repo", async () => {
+            jest.mocked(userRepository.getUserWithSecretsByEmail).mockResolvedValueOnce(USER_DBO);
+            await userService.getUserWithoutSecret(EMAIL);
+            expect(userRepository.getUserWithSecretsByEmail).toHaveBeenCalledWith(EMAIL);
+        });
+
+        it("return without secrets", async () => {
+            jest.mocked(userRepository.getUserWithSecretsByEmail).mockResolvedValueOnce(USER_DBO);
+            const actual = await userService.getUserWithoutSecret(EMAIL);
+            expect(actual).toMatchSnapshot();
+        });
+
+        it("throws not found if noe found", async () => {
+            jest.mocked(userRepository.getUserWithSecretsByEmail).mockResolvedValueOnce(null);
+            const test = () => userService.getUserWithoutSecret(EMAIL);
+            await expect(test).rejects.toMatchInlineSnapshot(`[Error: User not found]`);
         });
     });
 });